# docs/openapi/aisaas-v2.yaml
openapi: 3.1.0
info:
  title: ykt-aisaas V2 API
  description: |
    ykt-aisaas V2 API 规范，覆盖 Memory、Persona、Session、API Key 生命周期与 Audit 模块。

    ## 鉴权方式
    - `/api/v1/*` — API Key Bearer 鉴权（scope 校验）
    - `/internal/api/v1/*` — X-Internal-Token + loopback IP 鉴权（仅限内部超级租户）

    ## 错误响应
    - **SaaS 自有接口**（/api/v1/*）：统一信封 `{ code, message, requestId, data }`
    - **OpenAI 兼容接口**（/v1/*）：`{ error: { message, type, code }, requestId }`
  version: 2.0.0
  contact:
    name: ykt-aisaas Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8090
    description: Local development
  - url: https://api.aisaas.ykt.cn
    description: Production
...
paths:
...
  /api/v1/sessions/{deviceId}:
    post:
      tags: [session]
      operationId: createSession
      summary: 创建会话（含配额快照借记）
      description: |
        创建设备会话并预借配额（quota_snapshot）。
        Q8 决策：创建时从设备配额中借记 quotaInitial，返回借记后设备配额快照。
        dimension 指定借记维度（如 llm_tokens_in），会话结束后通过 /sessions/{sessionId}/end 结算实际消耗并退款。
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: 会话创建成功（含配额快照）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/QuotaExceeded'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/sessions/{sessionId}/end:
    post:
      tags: [session]
      operationId: endSession
      summary: 结算会话实际消耗（含配额退款）
      description: |
        结算会话实际消耗并退还剩余配额。
        Q8 决策：actualCost 包含实际消耗字段，quotaRefunded 为应退金额。
        status 为 failed 时标记会话异常结束，消耗按实际计入。
      security:
        - ApiKeyAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: 会话全局唯一 ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndSessionRequest'
      responses:
        '200':
          description: 会话已结算
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================
  # Internal APIKey 模块（2 个接口，Q6 决策）
  # ============================================================
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API-Key
      description: |
        SaaS 自有接口鉴权。调用方在 Authorization header 中传入 `Bearer <api_key>`。
        API Key 需具有对应 scope 权限。
      x-scope: required

    InternalTokenAuth:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: |
        内部超级租户接口鉴权。
        仅允许来自 loopback IP（127.0.0.0/8, ::1）的请求。
        Token 由平台运营后台生成，按租户隔离。

    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户门户（/portal/api/v1/*）JWT 鉴权

  # ============================================================
  # Parameters
  # ============================================================
  parameters:
    deviceId:
      name: deviceId
      in: path
      required: true
      description: 设备唯一标识（MAC 地址哈希或平台分配 ID）
      schema:
        type: string
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 错误码（与 errs.ErrorCode 对应）
          example: 40001
        message:
          type: string
          description: 错误描述
          example: 参数错误
        requestId:
          type: string
          description: 请求追踪 ID
          example: req_a1b2c3d4
        data:
          type: object
          description: 详细错误信息（可选）
...
    CreateSessionRequest:
      type: object
      required: [dimension, quotaInitial]
      properties:
        deviceId:
          type: string
          description: 设备 ID（路径参数同步）
        dimension:
          type: string
          enum: [llm_tokens_in, llm_tokens_out, tts_chars, asr_seconds]
          description: 借记维度
          example: llm_tokens_in
        quotaInitial:
          type: integer
          description: 申请借记量
          minimum: 1
          example: 2048
        personaId:
          type: integer
          format: int64
          description: 关联的 Persona ID（可选）
        metadata:
          type: object
          properties:
            ip:
              type: string
              description: 客户端 IP
            userAgent:
              type: string
            dimension:
              type: string
    CreateSessionResponse:
      type: object
      required: [sessionId, deviceId, dimension, quotaInitial, quotaRemaining, quotaSnapshot, createdAt]
      properties:
        sessionId:
          type: string
          format: uuid
          description: 会话 ID
        deviceId:
          type: string
        dimension:
          type: string
          enum: [llm_tokens_in, llm_tokens_out, tts_chars, asr_seconds]
        quotaInitial:
          type: integer
          description: 申请借记量
        quotaRemaining:
          type: integer
          description: 借记后剩余配额
        quotaSnapshot:
          $ref: '#/components/schemas/QuotaSnapshot'
        createdAt:
          type: string
          format: date-time
    QuotaSnapshot:
      type: object
      required: [tenantId, deviceQuotaLimit, deviceQuotaUsed, snapshotTime]
      description: Q8 决策：设备配额快照（创建会话时借记快照）
      properties:
        tenantId:
          type: integer
          format: int64
          description: 所属租户 ID
        deviceQuotaLimit:
          type: integer
          description: 设备配额上限
        deviceQuotaUsed:
          type: integer
          description: 设备已使用量
        deviceQuotaRemaining:
          type: integer
          description: 设备剩余配额
        snapshotTime:
          type: string
          format: date-time
...
    EndSessionRequest:
      type: object
      required: [actualCost, status]
      properties:
        actualCost:
          $ref: '#/components/schemas/QuotaUsage'
        status:
          type: string
          enum: [success, failed]
          description: 会话结束状态
    EndSessionResponse:
      type: object
      required: [sessionId, quotaUsed, quotaRefunded, endedAt]
      properties:
        sessionId:
          type: string
          format: uuid
        quotaUsed:
          $ref: '#/components/schemas/QuotaUsage'
        quotaRefunded:
          $ref: '#/components/schemas/QuotaUsage'
          description: 实际退款量（quotaInitial - quotaUsed）
        endedAt:
          type: string
          format: date-time
    QuotaUsage:
      type: object
      description: 配额使用量（支持多维度）
      properties:
        llm_tokens_in:
          type: integer
        llm_tokens_out:
          type: integer
        tts_chars:
          type: integer
        asr_seconds:
          type: integer
      example:
        llm_tokens_in: 1500
        llm_tokens_out: 800

    # ----- API Key 模块 -----
...
  responses:
    BadRequest:
      description: 参数错误或请求体解析失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40001
            message: 参数错误
            requestId: req_a1b2c3d4
    Unauthorized:
      description: 认证失败（API Key 无效或已过期）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40101
            message: API Key 无效
            requestId: req_a1b2c3d4
    QuotaExceeded:
      description: 配额不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40201
            message: 配额不足
            requestId: req_a1b2c3d4
...
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40302
            message: 权限不足
            requestId: req_a1b2c3d4
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40404
            message: 资源不存在
            requestId: req_a1b2c3d4
    Conflict:
      description: 资源已存在或状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40901
            message: 资源已存在
            requestId: req_a1b2c3d4
    InternalError:
      description: 系统内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 50002
            message: 系统内部错误
            requestId: req_a1b2c3d4